Use central platform checks and set font on non-ubuntu linux
